Move StreamEstimator to utils, changelog
