+++ CHANGELOG.rst
@@ -13,7 +13,7 @@ Added
 
 - Added `filters` as a replacement for `stat`. (#77)
 
-- Added `StreamEstimator` to estimate the remaining length of a stream. (#79)
+- Added `utils.StreamEstimator` to estimate the remaining length of a stream. (#79, #91)
 
 - Added `utils.stream_groupby`: Split a stream into sub-streams by key.
